+++ jason/src/peer/mod.rs
@@ -3,13 +3,12 @@
 //! [1]: https://www.w3.org/TR/webrtc/#rtcpeerconnection-interface
 
 mod ice_server;
+mod media_connections;
 mod peer_con;
 mod repo;
 
 #[doc(inline)]
 pub use self::{
-    peer_con::{
-        Id as PeerId, PeerConnection, PeerEvent, PeerEventHandler, Sdp,
-    },
+    peer_con::{Id as PeerId, PeerConnection, PeerEvent, PeerEventHandler},
     repo::PeerRepository,
 };
